Extract compilation logic to a `@rule_helper`.
[ci skip-rust]
[ci skip-build-wheels]
